chore: extract `CoFieldInit` type into its own file
