+++ packages/jazz-tools/src/tools/internal.ts
@@ -1,5 +1,6 @@
 export * from "./coValues/interfaces.js";
 export * from "./coValues/CoValueBase.js";
+export * from "./coValues/CoFieldInit.js";
 export * from "./implementation/inspect.js";
 export * from "./implementation/symbols.js";
